requirements.txt
alembic==1.3.2
blinker==1.4
Click==7.0
Flask==1.1.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2==2.8.4
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
six==1.13.0
SQLAlchemy==1.3.12
Werkzeug==0.16.0
